Cleanup with new config after rebase, remove CRLF, change paths to texts in CardFront svg. Style with less instead of inline
