--- src/profile/components/AccessCard/CardFront.tsx
@@ -1,28 +1,0 @@
-import React from 'react';
-
-export interface IProps {
-  name: string;
-  studentNumber: string;
-}
-
-const CardFront = ({ name, studentNumber }: IProps) => (
-  <svg viewBox="0 0 216 344">
-    <rect width="216" height="344" rx="15" fill="#A6B47E"/>
-    <path d="M0 288H216V329C216 337.284 209.284 344 201 344H15C6.71573 344 0 337.284 0 329V288Z" fill="#394299"/>
-    <rect x="5" y="246" width="206" height="2" fill="#3C3A3A"/>
-    <rect x="68" y="98" width="80" height="108" fill="#C4C4C4"/>
-    <path d="M12.5527 224.01C12.5527 224.846 12.4121 225.576 12.1309 226.201C11.8496 226.822 11.4512 227.297 10.9355 227.625C10.4199 227.953 9.81836 228.117 9.13086 228.117C8.45898 228.117 7.86328 227.953 7.34375 227.625C6.82422 227.293 6.41992 226.822 6.13086 226.213C5.8457 225.6 5.69922 224.891 5.69141 224.086V223.471C5.69141 222.65 5.83398 221.926 6.11914 221.297C6.4043 220.668 6.80664 220.188 7.32617 219.855C7.84961 219.52 8.44727 219.352 9.11914 219.352C9.80273 219.352 10.4043 219.518 10.9238 219.85C11.4473 220.178 11.8496 220.656 12.1309 221.285C12.4121 221.91 12.5527 222.639 12.5527 223.471V224.01ZM11.4336 223.459C11.4336 222.447 11.2305 221.672 10.8242 221.133C10.418 220.59 9.84961 220.318 9.11914 220.318C8.4082 220.318 7.84766 220.59 7.4375 221.133C7.03125 221.672 6.82227 222.422 6.81055 223.383V224.01C6.81055 224.99 7.01562 225.762 7.42578 226.324C7.83984 226.883 8.4082 227.162 9.13086 227.162C9.85742 227.162 10.4199 226.898 10.8184 226.371C11.2168 225.84 11.4219 225.08 11.4336 224.092V223.459ZM15.248 228H14.1641V219H15.248V228ZM19.6191 228.117C18.7598 228.117 18.0605 227.836 17.5215 227.273C16.9824 226.707 16.7129 225.951 16.7129 225.006V224.807C16.7129 224.178 16.832 223.617 17.0703 223.125C17.3125 222.629 17.6484 222.242 18.0781 221.965C18.5117 221.684 18.9805 221.543 19.4844 221.543C20.3086 221.543 20.9492 221.814 21.4062 222.357C21.8633 222.9 22.0918 223.678 22.0918 224.689V225.141H17.7969C17.8125 225.766 17.9941 226.271 18.3418 226.658C18.6934 227.041 19.1387 227.232 19.6777 227.232C20.0605 227.232 20.3848 227.154 20.6504 226.998C20.916 226.842 21.1484 226.635 21.3477 226.377L22.0098 226.893C21.4785 227.709 20.6816 228.117 19.6191 228.117ZM19.4844 222.434C19.0469 222.434 18.6797 222.594 18.3828 222.914C18.0859 223.23 17.9023 223.676 17.832 224.25H21.0078V224.168C20.9766 223.617 20.8281 223.191 20.5625 222.891C20.2969 222.586 19.9375 222.434 19.4844 222.434ZM31.209 225.773H27.6348L26.832 228H25.6719L28.9297 219.469H29.9141L33.1777 228H32.0234L31.209 225.773ZM27.9746 224.848H30.875L29.4219 220.857L27.9746 224.848ZM35.1816 221.66L35.2168 222.457C35.7012 221.848 36.334 221.543 37.1152 221.543C38.4551 221.543 39.1309 222.299 39.1426 223.811V228H38.0586V223.805C38.0547 223.348 37.9492 223.01 37.7422 222.791C37.5391 222.572 37.2207 222.463 36.7871 222.463C36.4355 222.463 36.127 222.557 35.8613 222.744C35.5957 222.932 35.3887 223.178 35.2402 223.482V228H34.1562V221.66H35.1816ZM40.5137 224.777C40.5137 223.805 40.7441 223.023 41.2051 222.434C41.666 221.84 42.2695 221.543 43.0156 221.543C43.7578 221.543 44.3457 221.797 44.7793 222.305V219H45.8633V228H44.8672L44.8145 227.32C44.3809 227.852 43.7773 228.117 43.0039 228.117C42.2695 228.117 41.6699 227.816 41.2051 227.215C40.7441 226.613 40.5137 225.828 40.5137 224.859V224.777ZM41.5977 224.9C41.5977 225.619 41.7461 226.182 42.043 226.588C42.3398 226.994 42.75 227.197 43.2734 227.197C43.9609 227.197 44.4629 226.889 44.7793 226.271V223.359C44.4551 222.762 43.957 222.463 43.2852 222.463C42.7539 222.463 42.3398 222.668 42.043 223.078C41.7461 223.488 41.5977 224.096 41.5977 224.9ZM50.1816 228.117C49.3223 228.117 48.623 227.836 48.084 227.273C47.5449 226.707 47.2754 225.951 47.2754 225.006V224.807C47.2754 224.178 47.3945 223.617 47.6328 223.125C47.875 222.629 48.2109 222.242 48.6406 221.965C49.0742 221.684 49.543 221.543 50.0469 221.543C50.8711 221.543 51.5117 221.814 51.9688 222.357C52.4258 222.9 52.6543 223.678 52.6543 224.689V225.141H48.3594C48.375 225.766 48.5566 226.271 48.9043 226.658C49.2559 227.041 49.7012 227.232 50.2402 227.232C50.623 227.232 50.9473 227.154 51.2129 226.998C51.4785 226.842 51.7109 226.635 51.9102 226.377L52.5723 226.893C52.041 227.709 51.2441 228.117 50.1816 228.117ZM50.0469 222.434C49.6094 222.434 49.2422 222.594 48.9453 222.914C48.6484 223.23 48.4648 223.676 48.3945 224.25H51.5703V224.168C51.5391 223.617 51.3906 223.191 51.125 222.891C50.8594 222.586 50.5 222.434 50.0469 222.434ZM56.9785 222.633C56.8145 222.605 56.6367 222.592 56.4453 222.592C55.7344 222.592 55.252 222.895 54.998 223.5V228H53.9141V221.66H54.9688L54.9863 222.393C55.3418 221.826 55.8457 221.543 56.498 221.543C56.709 221.543 56.8691 221.57 56.9785 221.625V222.633ZM61.6719 226.318C61.6719 226.025 61.5605 225.799 61.3379 225.639C61.1191 225.475 60.7344 225.334 60.1836 225.217C59.6367 225.1 59.2012 224.959 58.877 224.795C58.5566 224.631 58.3184 224.436 58.1621 224.209C58.0098 223.982 57.9336 223.713 57.9336 223.4C57.9336 222.881 58.1523 222.441 58.5898 222.082C59.0312 221.723 59.5938 221.543 60.2773 221.543C60.9961 221.543 61.5781 221.729 62.0234 222.1C62.4727 222.471 62.6973 222.945 62.6973 223.523H61.6074C61.6074 223.227 61.4805 222.971 61.2266 222.756C60.9766 222.541 60.6602 222.434 60.2773 222.434C59.8828 222.434 59.5742 222.52 59.3516 222.691C59.1289 222.863 59.0176 223.088 59.0176 223.365C59.0176 223.627 59.1211 223.824 59.3281 223.957C59.5352 224.09 59.9082 224.217 60.4473 224.338C60.9902 224.459 61.4297 224.604 61.7656 224.771C62.1016 224.939 62.3496 225.143 62.5098 225.381C62.6738 225.615 62.7559 225.902 62.7559 226.242C62.7559 226.809 62.5293 227.264 62.0762 227.607C61.623 227.947 61.0352 228.117 60.3125 228.117C59.8047 228.117 59.3555 228.027 58.9648 227.848C58.5742 227.668 58.2676 227.418 58.0449 227.098C57.8262 226.773 57.7168 226.424 57.7168 226.049H58.8008C58.8203 226.412 58.9648 226.701 59.2344 226.916C59.5078 227.127 59.8672 227.232 60.3125 227.232C60.7227 227.232 61.0508 227.15 61.2969 226.986C61.5469 226.818 61.6719 226.596 61.6719 226.318ZM8.50391 238.197C7.53906 237.92 6.83594 237.58 6.39453 237.178C5.95703 236.771 5.73828 236.271 5.73828 235.678C5.73828 235.006 6.00586 234.451 6.54102 234.014C7.08008 233.572 7.7793 233.352 8.63867 233.352C9.22461 233.352 9.74609 233.465 10.2031 233.691C10.6641 233.918 11.0195 234.23 11.2695 234.629C11.5234 235.027 11.6504 235.463 11.6504 235.936H10.5195C10.5195 235.42 10.3555 235.016 10.0273 234.723C9.69922 234.426 9.23633 234.277 8.63867 234.277C8.08398 234.277 7.65039 234.4 7.33789 234.646C7.0293 234.889 6.875 235.227 6.875 235.66C6.875 236.008 7.02148 236.303 7.31445 236.545C7.61133 236.783 8.11328 237.002 8.82031 237.201C9.53125 237.4 10.0859 237.621 10.4844 237.863C10.8867 238.102 11.1836 238.381 11.375 238.701C11.5703 239.021 11.668 239.398 11.668 239.832C11.668 240.523 11.3984 241.078 10.8594 241.496C10.3203 241.91 9.59961 242.117 8.69727 242.117C8.11133 242.117 7.56445 242.006 7.05664 241.783C6.54883 241.557 6.15625 241.248 5.87891 240.857C5.60547 240.467 5.46875 240.023 5.46875 239.527H6.59961C6.59961 240.043 6.78906 240.451 7.16797 240.752C7.55078 241.049 8.06055 241.197 8.69727 241.197C9.29102 241.197 9.74609 241.076 10.0625 240.834C10.3789 240.592 10.5371 240.262 10.5371 239.844C10.5371 239.426 10.3906 239.104 10.0977 238.877C9.80469 238.646 9.27344 238.42 8.50391 238.197ZM14.416 234.125V235.66H15.5996V236.498H14.416V240.43C14.416 240.684 14.4688 240.875 14.5742 241.004C14.6797 241.129 14.8594 241.191 15.1133 241.191C15.2383 241.191 15.4102 241.168 15.6289 241.121V242C15.3438 242.078 15.0664 242.117 14.7969 242.117C14.3125 242.117 13.9473 241.971 13.7012 241.678C13.4551 241.385 13.332 240.969 13.332 240.43V236.498H12.1777V235.66H13.332V234.125H14.416ZM16.4668 238.771C16.4668 238.15 16.5879 237.592 16.8301 237.096C17.0762 236.6 17.416 236.217 17.8496 235.947C18.2871 235.678 18.7852 235.543 19.3438 235.543C20.207 235.543 20.9043 235.842 21.4355 236.439C21.9707 237.037 22.2383 237.832 22.2383 238.824V238.9C22.2383 239.518 22.1191 240.072 21.8809 240.564C21.6465 241.053 21.3086 241.434 20.8672 241.707C20.4297 241.98 19.9258 242.117 19.3555 242.117C18.4961 242.117 17.7988 241.818 17.2637 241.221C16.7324 240.623 16.4668 239.832 16.4668 238.848V238.771ZM17.5566 238.9C17.5566 239.604 17.7188 240.168 18.043 240.594C18.3711 241.02 18.8086 241.232 19.3555 241.232C19.9062 241.232 20.3438 241.018 20.668 240.588C20.9922 240.154 21.1543 239.549 21.1543 238.771C21.1543 238.076 20.9883 237.514 20.6562 237.084C20.3281 236.65 19.8906 236.434 19.3438 236.434C18.8086 236.434 18.377 236.646 18.0488 237.072C17.7207 237.498 17.5566 238.107 17.5566 238.9ZM25.3672 239.064L24.6875 239.773V242H23.6035V233H24.6875V238.443L25.2676 237.746L27.2422 235.66H28.5605L26.0938 238.309L28.8477 242H27.5762L25.3672 239.064ZM31.4492 239.064L30.7695 239.773V242H29.6855V233H30.7695V238.443L31.3496 237.746L33.3242 235.66H34.6426L32.1758 238.309L34.9297 242H33.6582L31.4492 239.064ZM38.2754 242.117C37.416 242.117 36.7168 241.836 36.1777 241.273C35.6387 240.707 35.3691 239.951 35.3691 239.006V238.807C35.3691 238.178 35.4883 237.617 35.7266 237.125C35.9688 236.629 36.3047 236.242 36.7344 235.965C37.168 235.684 37.6367 235.543 38.1406 235.543C38.9648 235.543 39.6055 235.814 40.0625 236.357C40.5195 236.9 40.748 237.678 40.748 238.689V239.141H36.4531C36.4688 239.766 36.6504 240.271 36.998 240.658C37.3496 241.041 37.7949 241.232 38.334 241.232C38.7168 241.232 39.041 241.154 39.3066 240.998C39.5723 240.842 39.8047 240.635 40.0039 240.377L40.666 240.893C40.1348 241.709 39.3379 242.117 38.2754 242.117ZM38.1406 236.434C37.7031 236.434 37.3359 236.594 37.0391 236.914C36.7422 237.23 36.5586 237.676 36.4883 238.25H39.6641V238.168C39.6328 237.617 39.4844 237.191 39.2188 236.891C38.9531 236.586 38.5938 236.434 38.1406 236.434ZM45.0723 236.633C44.9082 236.605 44.7305 236.592 44.5391 236.592C43.8281 236.592 43.3457 236.895 43.0918 237.5V242H42.0078V235.66H43.0625L43.0801 236.393C43.4355 235.826 43.9395 235.543 44.5918 235.543C44.8027 235.543 44.9629 235.57 45.0723 235.625V236.633Z" fill="#3C3A3A"/>
-    <path d="M11.2168 276.078L7.68359 284H6.54688L10.0684 276.359H5.45117V275.469H11.2168V276.078ZM16.7012 275.463V276.383H16.502C15.6582 276.398 14.9863 276.648 14.4863 277.133C13.9863 277.617 13.6973 278.299 13.6191 279.178C14.0684 278.662 14.6816 278.404 15.459 278.404C16.2012 278.404 16.793 278.666 17.2344 279.189C17.6797 279.713 17.9023 280.389 17.9023 281.217C17.9023 282.096 17.6621 282.799 17.1816 283.326C16.7051 283.854 16.0645 284.117 15.2598 284.117C14.4434 284.117 13.7812 283.805 13.2734 283.18C12.7656 282.551 12.5117 281.742 12.5117 280.754V280.338C12.5117 278.768 12.8457 277.568 13.5137 276.74C14.1855 275.908 15.1836 275.482 16.5078 275.463H16.7012ZM15.2773 279.307C14.9062 279.307 14.5645 279.418 14.252 279.641C13.9395 279.863 13.7227 280.143 13.6016 280.479V280.877C13.6016 281.58 13.7598 282.146 14.0762 282.576C14.3926 283.006 14.7871 283.221 15.2598 283.221C15.748 283.221 16.1309 283.041 16.4082 282.682C16.6895 282.322 16.8301 281.852 16.8301 281.27C16.8301 280.684 16.6875 280.211 16.4023 279.852C16.1211 279.488 15.7461 279.307 15.2773 279.307ZM23.4395 275.463V276.383H23.2402C22.3965 276.398 21.7246 276.648 21.2246 277.133C20.7246 277.617 20.4355 278.299 20.3574 279.178C20.8066 278.662 21.4199 278.404 22.1973 278.404C22.9395 278.404 23.5312 278.666 23.9727 279.189C24.418 279.713 24.6406 280.389 24.6406 281.217C24.6406 282.096 24.4004 282.799 23.9199 283.326C23.4434 283.854 22.8027 284.117 21.998 284.117C21.1816 284.117 20.5195 283.805 20.0117 283.18C19.5039 282.551 19.25 281.742 19.25 280.754V280.338C19.25 278.768 19.584 277.568 20.252 276.74C20.9238 275.908 21.9219 275.482 23.2461 275.463H23.4395ZM22.0156 279.307C21.6445 279.307 21.3027 279.418 20.9902 279.641C20.6777 279.863 20.4609 280.143 20.3398 280.479V280.877C20.3398 281.58 20.498 282.146 20.8145 282.576C21.1309 283.006 21.5254 283.221 21.998 283.221C22.4863 283.221 22.8691 283.041 23.1465 282.682C23.4277 282.322 23.5684 281.852 23.5684 281.27C23.5684 280.684 23.4258 280.211 23.1406 279.852C22.8594 279.488 22.4844 279.307 22.0156 279.307ZM31.0977 277.695C31.0977 278.121 30.9844 278.5 30.7578 278.832C30.5352 279.164 30.2324 279.424 29.8496 279.611C30.2949 279.803 30.6465 280.082 30.9043 280.449C31.166 280.816 31.2969 281.232 31.2969 281.697C31.2969 282.436 31.0469 283.023 30.5469 283.461C30.0508 283.898 29.3965 284.117 28.584 284.117C27.7637 284.117 27.1055 283.898 26.6094 283.461C26.1172 283.02 25.8711 282.432 25.8711 281.697C25.8711 281.236 25.9961 280.82 26.2461 280.449C26.5 280.078 26.8496 279.797 27.2949 279.605C26.916 279.418 26.6172 279.158 26.3984 278.826C26.1797 278.494 26.0703 278.117 26.0703 277.695C26.0703 276.977 26.3008 276.406 26.7617 275.984C27.2227 275.562 27.8301 275.352 28.584 275.352C29.334 275.352 29.9395 275.562 30.4004 275.984C30.8652 276.406 31.0977 276.977 31.0977 277.695ZM30.2129 281.674C30.2129 281.197 30.0605 280.809 29.7559 280.508C29.4551 280.207 29.0605 280.057 28.5723 280.057C28.084 280.057 27.6914 280.205 27.3945 280.502C27.1016 280.799 26.9551 281.189 26.9551 281.674C26.9551 282.158 27.0977 282.539 27.3828 282.816C27.6719 283.094 28.0723 283.232 28.584 283.232C29.0918 283.232 29.4902 283.094 29.7793 282.816C30.0684 282.535 30.2129 282.154 30.2129 281.674ZM28.584 276.242C28.1582 276.242 27.8125 276.375 27.5469 276.641C27.2852 276.902 27.1543 277.26 27.1543 277.713C27.1543 278.146 27.2832 278.498 27.541 278.768C27.8027 279.033 28.1504 279.166 28.584 279.166C29.0176 279.166 29.3633 279.033 29.6211 278.768C29.8828 278.498 30.0137 278.146 30.0137 277.713C30.0137 277.279 29.8789 276.926 29.6094 276.652C29.3398 276.379 28.998 276.242 28.584 276.242ZM33.1602 279.723L33.5938 275.469H37.9648V276.471H34.5137L34.2559 278.797C34.6738 278.551 35.1484 278.428 35.6797 278.428C36.457 278.428 37.0742 278.686 37.5312 279.201C37.9883 279.713 38.2168 280.406 38.2168 281.281C38.2168 282.16 37.9785 282.854 37.502 283.361C37.0293 283.865 36.3672 284.117 35.5156 284.117C34.7617 284.117 34.1465 283.908 33.6699 283.49C33.1934 283.072 32.9219 282.494 32.8555 281.756H33.8809C33.9473 282.244 34.1211 282.613 34.4023 282.863C34.6836 283.109 35.0547 283.232 35.5156 283.232C36.0195 283.232 36.4141 283.061 36.6992 282.717C36.9883 282.373 37.1328 281.898 37.1328 281.293C37.1328 280.723 36.9766 280.266 36.6641 279.922C36.3555 279.574 35.9434 279.4 35.4277 279.4C34.9551 279.4 34.584 279.504 34.3145 279.711L34.0273 279.945L33.1602 279.723ZM43.9766 281.135H45.1602V282.02H43.9766V284H42.8867V282.02H39.002V281.381L42.8223 275.469H43.9766V281.135ZM40.2324 281.135H42.8867V276.951L42.7578 277.186L40.2324 281.135Z" fill="#3C3A3A"/>
-    <path d="M186.523 193.629C185.859 193.629 185.352 193.863 185 194.332C184.641 194.801 184.266 195.645 183.875 196.863C183.477 198.082 183.086 199.047 182.703 199.758C181.656 201.695 180.246 202.664 178.473 202.664C177.551 202.664 176.73 202.406 176.012 201.891C175.285 201.367 174.719 200.621 174.312 199.652C173.906 198.676 173.703 197.582 173.703 196.371C173.703 195.152 173.926 194.066 174.371 193.113C174.809 192.16 175.43 191.422 176.234 190.898C177.039 190.367 177.953 190.102 178.977 190.102V193.617C178.195 193.617 177.59 193.863 177.16 194.355C176.723 194.848 176.504 195.539 176.504 196.43C176.504 197.289 176.688 197.957 177.055 198.434C177.414 198.91 177.891 199.148 178.484 199.148C179.039 199.148 179.504 198.871 179.879 198.316C180.254 197.754 180.605 196.93 180.934 195.844C181.535 193.844 182.281 192.387 183.172 191.473C184.062 190.559 185.172 190.102 186.5 190.102C187.977 190.102 189.137 190.66 189.98 191.777C190.816 192.895 191.234 194.398 191.234 196.289C191.234 197.602 190.996 198.797 190.52 199.875C190.035 200.953 189.375 201.777 188.539 202.348C187.703 202.91 186.734 203.191 185.633 203.191V199.664C187.516 199.664 188.457 198.539 188.457 196.289C188.457 195.453 188.289 194.801 187.953 194.332C187.609 193.863 187.133 193.629 186.523 193.629ZM176.785 174.867V180.094H191V183.609H176.785V188.766H173.938V174.867H176.785ZM173.938 159.938H185.176C187.043 159.938 188.52 160.523 189.605 161.695C190.691 162.859 191.234 164.453 191.234 166.477C191.234 168.469 190.707 170.051 189.652 171.223C188.598 172.395 187.148 172.992 185.305 173.016H173.938V169.5H185.199C186.316 169.5 187.133 169.234 187.648 168.703C188.156 168.164 188.41 167.422 188.41 166.477C188.41 164.5 187.371 163.496 185.293 163.465H173.938V159.938ZM191 157.055H173.938V151.805C173.938 150.305 174.277 148.965 174.957 147.785C175.629 146.598 176.59 145.672 177.84 145.008C179.082 144.344 180.496 144.012 182.082 144.012H182.867C184.453 144.012 185.863 144.34 187.098 144.996C188.332 145.645 189.289 146.562 189.969 147.75C190.648 148.938 190.992 150.277 191 151.77V157.055ZM176.785 153.539H188.176V151.84C188.176 150.465 187.727 149.414 186.828 148.688C185.93 147.961 184.645 147.59 182.973 147.574H182.07C180.336 147.574 179.023 147.934 178.133 148.652C177.234 149.371 176.785 150.422 176.785 151.805V153.539ZM183.605 131.18V137.93H188.176V130.008H191V141.445H173.938V130.031H176.785V137.93H180.852V131.18H183.605ZM191 114.07V117.586L179.773 124.43H191V127.945H173.938V124.43L185.188 117.574H173.938V114.07H191ZM176.785 98.5078V103.734H191V107.25H176.785V112.406H173.938V98.5078H176.785Z" fill="#3C3A3A"/>
-    <path d="M7.33594 267.465C6.69271 267.28 6.22396 267.053 5.92969 266.785C5.63802 266.514 5.49219 266.181 5.49219 265.785C5.49219 265.337 5.67057 264.967 6.02734 264.676C6.38672 264.382 6.85286 264.234 7.42578 264.234C7.81641 264.234 8.16406 264.31 8.46875 264.461C8.77604 264.612 9.01302 264.82 9.17969 265.086C9.34896 265.352 9.43359 265.642 9.43359 265.957H8.67969C8.67969 265.613 8.57031 265.344 8.35156 265.148C8.13281 264.951 7.82422 264.852 7.42578 264.852C7.05599 264.852 6.76693 264.934 6.55859 265.098C6.35286 265.259 6.25 265.484 6.25 265.773C6.25 266.005 6.34766 266.202 6.54297 266.363C6.74089 266.522 7.07552 266.668 7.54688 266.801C8.02083 266.934 8.39062 267.081 8.65625 267.242C8.92448 267.401 9.1224 267.587 9.25 267.801C9.38021 268.014 9.44531 268.266 9.44531 268.555C9.44531 269.016 9.26562 269.385 8.90625 269.664C8.54688 269.94 8.06641 270.078 7.46484 270.078C7.07422 270.078 6.70964 270.004 6.37109 269.855C6.03255 269.704 5.77083 269.499 5.58594 269.238C5.40365 268.978 5.3125 268.682 5.3125 268.352H6.06641C6.06641 268.695 6.19271 268.967 6.44531 269.168C6.70052 269.366 7.04036 269.465 7.46484 269.465C7.86068 269.465 8.16406 269.384 8.375 269.223C8.58594 269.061 8.69141 268.841 8.69141 268.562C8.69141 268.284 8.59375 268.069 8.39844 267.918C8.20312 267.764 7.84896 267.613 7.33594 267.465ZM14.3398 264.93H12.5117V270H11.7656V264.93H9.94141V264.312H14.3398V264.93ZM19.1875 264.312V268.18C19.1849 268.716 19.0156 269.155 18.6797 269.496C18.3464 269.837 17.8932 270.029 17.3203 270.07L17.1211 270.078C16.4987 270.078 16.0026 269.91 15.6328 269.574C15.263 269.238 15.0755 268.776 15.0703 268.188V264.312H15.8125V268.164C15.8125 268.576 15.9258 268.896 16.1523 269.125C16.3789 269.352 16.7018 269.465 17.1211 269.465C17.5456 269.465 17.8698 269.352 18.0938 269.125C18.3203 268.898 18.4336 268.579 18.4336 268.168V264.312H19.1875ZM20.3711 270V264.312H21.9766C22.4714 264.312 22.9089 264.422 23.2891 264.641C23.6693 264.859 23.9622 265.171 24.168 265.574C24.3763 265.978 24.4818 266.441 24.4844 266.965V267.328C24.4844 267.865 24.3802 268.335 24.1719 268.738C23.9661 269.142 23.6706 269.452 23.2852 269.668C22.9023 269.884 22.4557 269.995 21.9453 270H20.3711ZM21.1211 264.93V269.387H21.9102C22.4883 269.387 22.9375 269.207 23.2578 268.848C23.5807 268.488 23.7422 267.977 23.7422 267.312V266.98C23.7422 266.335 23.5898 265.833 23.2852 265.477C22.9831 265.117 22.5534 264.935 21.9961 264.93H21.1211ZM28.8359 267.371H26.3711V269.387H29.2344V270H25.6211V264.312H29.1953V264.93H26.3711V266.758H28.8359V267.371ZM34.5391 270H33.7852L30.9219 265.617V270H30.168V264.312H30.9219L33.793 268.715V264.312H34.5391V270ZM39.6836 264.93H37.8555V270H37.1094V264.93H35.2852V264.312H39.6836V264.93ZM44.8984 270H44.1445L41.2812 265.617V270H40.5273V264.312H41.2812L44.1523 268.715V264.312H44.8984V270ZM48.3164 267.699H46.9805V270H46.2266V264.312H48.1094C48.75 264.312 49.2422 264.458 49.5859 264.75C49.9323 265.042 50.1055 265.466 50.1055 266.023C50.1055 266.378 50.0091 266.686 49.8164 266.949C49.6263 267.212 49.3607 267.409 49.0195 267.539L50.3555 269.953V270H49.5508L48.3164 267.699ZM46.9805 267.086H48.1328C48.5052 267.086 48.8008 266.99 49.0195 266.797C49.2409 266.604 49.3516 266.346 49.3516 266.023C49.3516 265.672 49.2461 265.402 49.0352 265.215C48.8268 265.027 48.5247 264.932 48.1289 264.93H46.9805V267.086ZM51.0625 269.621C51.0625 269.496 51.099 269.392 51.1719 269.309C51.2474 269.225 51.3594 269.184 51.5078 269.184C51.6562 269.184 51.7682 269.225 51.8438 269.309C51.9219 269.392 51.9609 269.496 51.9609 269.621C51.9609 269.741 51.9219 269.841 51.8438 269.922C51.7682 270.003 51.6562 270.043 51.5078 270.043C51.3594 270.043 51.2474 270.003 51.1719 269.922C51.099 269.841 51.0625 269.741 51.0625 269.621ZM53.1328 269.621C53.1328 269.496 53.1693 269.392 53.2422 269.309C53.3177 269.225 53.4297 269.184 53.5781 269.184C53.7266 269.184 53.8385 269.225 53.9141 269.309C53.9922 269.392 54.0312 269.496 54.0312 269.621C54.0312 269.741 53.9922 269.841 53.9141 269.922C53.8385 270.003 53.7266 270.043 53.5781 270.043C53.4297 270.043 53.3177 270.003 53.2422 269.922C53.1693 269.841 53.1328 269.741 53.1328 269.621ZM53.1367 266.172C53.1367 266.047 53.1732 265.943 53.2461 265.859C53.3216 265.776 53.4336 265.734 53.582 265.734C53.7305 265.734 53.8424 265.776 53.918 265.859C53.9961 265.943 54.0352 266.047 54.0352 266.172C54.0352 266.292 53.9961 266.392 53.918 266.473C53.8424 266.553 53.7305 266.594 53.582 266.594C53.4336 266.594 53.3216 266.553 53.2461 266.473C53.1732 266.392 53.1367 266.292 53.1367 266.172Z" fill="#3C3A3A"/>
-    <rect x="19" y="19" width="178.56" height="48.672" fill="url(#pattern0)"/>
-    <rect x="12" y="299" width="90" height="4" fill="white"/>
-    <rect x="12" y="313" width="192" height="4" fill="white"/>
-    <rect x="12" y="327" width="192" height="4" fill="white"/>
-    <rect x="114" y="299" width="90" height="4" fill="white"/>
-    <circle cx="108" cy="301" r="2" fill="white"/>
-    <image x="20" y="20" id="image0" width="178" height="48" href="https://innsida.ntnu.no/documents/10157/2546401449/logo_ntnu_bokm-sample-a.png/435f2d98-f0d0-4db8-af20-4b77ba45e27a?t=1456603981519"/>
-  </svg>
-)
-
-export default CardFront;
