+++ src/profile/components/Settings/AccessCard/Barcode.tsx
